backend/requirements/base.txt
django>=6.0,<6.1
psycopg[binary]>=3.2
django-environ>=0.12
pillow>=11.0
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.7
dj-rest-auth>=7.0
django-allauth[socialaccount]>=65.0
drf-spectacular>=0.28
django-modeltranslation>=0.20
markdown>=3.7
bleach>=6.2
django-autoslug>=1.9
channels[daphne]>=4.2
channels-redis>=4.2
